updates package.json and rollup.config.js
