+++ package.json
@@ -1,51 +1,27 @@
 {
-  "name": "rally-study-01",
-  "version": "0.2.0",
+  "name": "rally-web-platform",
+  "version": "0.1.0",
   "scripts": {
-    "build": "npm run build:schema && npm run build:addon",
-    "build:addon": "npm run build:glean && rollup -c && web-ext build --overwrite-dest",
-    "build:glean": "npm run build:glean:metrics && npm run build:glean:docs",
-    "build:glean:metrics": "glean translate ./metrics.yaml ./pings.yaml -f javascript -o src/generated",
-    "build:glean:docs": "glean translate ./metrics.yaml ./pings.yaml -f markdown -o docs",
-    "build:schema": "node schemas/generate-schema.mjs",
+    "build": "rollup -c",
     "build:storybook": "build-storybook -s ./",
-    "build:web": "rollup -c rollup.config.web.js",
-    "dev": "rollup -c --config-enable-developer-mode",
+    "dev": "rollup -c -w",
     "doc": "jsdoc -c jsdoc.conf.json && mkdir -p docs/images && cp images/rally-logo.png docs/images/",
-    "lint": "npm-run-all lint:web-ext lint:eslint",
-    "lint:eslint": "eslint .",
-    "lint:glean": "glean glinter ./metrics.yaml ./pings.yaml",
-    "lint:web-ext": "web-ext lint",
-    "package": "npm run build && web-ext build --overwrite-dest --ignore-files \"./tests/**/*\" --ignore-files \"./src/**/*\" --ignore-files \"./rollup.config.js\"",
-    "start": "npm run build && web-ext run --target=chromium",
+    "lint": "eslint .",
     "test:unit": "jest",
     "test:integration": "npm run package && mv web-ext-artifacts/*.zip web-ext-artifacts/study.xpi && mocha --timeout 30000 \"./tests/integration/*.js\"",
-    "watch": "npm-run-all --parallel watch:raw watch:bundled",
-    "watch:raw": "npm run build:schema && npm run dev -- -w",
-    "watch:web": "rollup -c rollup.config.web.js -w",
-    "start:web": "sirv public --no-clear",
-    "watch:bundled": "npm run build:schema && web-ext run --target=chromium --watch-file dist/background.js --watch-file dist/content-scripts/attention-collector.js --watch-file schemas/measurements.1.schema.json",
-    "gh-pages": "gh-pages --dist public"
+    "gh-pages": "gh-pages --dist public",
+    "start": "sirv public --no-clear"
   },
   "jest": {
     "verbose": true,
     "moduleFileExtensions": [
       "js",
       "json",
-      "jsx",
       "ts",
       "tsx",
-      "node"
+      "node",
+      "svelte"
     ],
-    "moduleNameMapper": {
-      "^@mozilla/glean/webext$": "<rootDir>/node_modules/@mozilla/glean/dist/webext/index/webext.js",
-      "^@mozilla/glean/webext/plugins/encryption$": "<rootDir>/node_modules/@mozilla/glean/dist/webext/plugins/encryption.js",
-      "^@mozilla/glean/webext/private/metrics/(.*)$": "<rootDir>/node_modules/@mozilla/glean/dist/webext/core/metrics/types/$1.js",
-      "^@mozilla/glean/webext/private/ping$": "<rootDir>/node_modules/@mozilla/glean/dist/webext/core/pings/ping_type.js",
-      "^jose/jwe/compact/encrypt$": "<rootDir>/node_modules/jose/dist/node/cjs/jwe/compact/encrypt.js",
-      "^jose/jwk/parse": "<rootDir>/node_modules/jose/dist/node/cjs/jwk/parse.js",
-      "^jose/jwk/thumbprint": "<rootDir>/node_modules/jose/dist/node/cjs/jwk/thumbprint.js"
-    },
     "testEnvironment": "./tests/jest-custom-environment.js",
     "transform": {
       ".+\\.(css|styl|less|sass|scss|png|jpg|ttf|woff|woff2)$": "jest-transform-stub",
@@ -53,17 +29,12 @@
     },
     "setupFiles": [
       "jest-webextension-mock"
-    ],
-    "transformIgnorePatterns": [
-      "/node_modules/@mozilla/rally.js"
     ]
   },
   "devDependencies": {
     "@babel/plugin-transform-runtime": "^7.13.10",
     "@babel/preset-env": "^7.13.12",
     "@mozilla-protocol/core": "^12.1.0",
-    "@mozilla/glean": "^0.15.0",
-    "@mozilla/rally": "0.6.0",
     "@rollup/plugin-commonjs": "^18.0.0",
     "@rollup/plugin-node-resolve": "^11.2.1",
     "@rollup/plugin-replace": "^2.4.1",
@@ -83,7 +54,6 @@
     "jest-webextension-mock": "^3.7.6",
     "jsdoc": "^3.6.6",
     "micromodal": "^0.4.6",
-    "mocha": "^8.1.3",
     "npm-run-all": "^4.1.5",
     "rollup": "^2.33.1",
     "rollup-plugin-copy": "^3.4.0",
@@ -94,9 +64,7 @@
     "selenium-webdriver": "^4.0.0-alpha.7",
     "sinon-chrome": "^3.0.1",
     "svelte": "^3.38.2",
-    "tidy-jsdoc": "^1.4.0",
-    "web-ext": "^6.0.0",
-    "webextension-polyfill": "^0.7.0"
+    "tidy-jsdoc": "^1.4.0"
   },
   "dependencies": {
     "firebase": "^9.0.0-beta.2",
@@ -105,17 +73,17 @@
   "engines": {
     "node": ">=14.0.0"
   },
-  "description": "An exploratory study on what sites people use and how long they use them.",
+  "description": "Repository for the Rally web platform",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/mozilla-rally/rally-study-01.git"
+    "url": "git+https://github.com/mozilla-rally/rally-web-platform.git"
   },
   "keywords": [],
   "author": "Mozilla Rally",
   "license": "MPL-2.0",
   "private": true,
   "bugs": {
-    "url": "https://github.com/mozilla-rally/rally-study-01/issues"
+    "url": "https://github.com/mozilla-rally/rally-web-platform/issues"
   },
-  "homepage": "https://github.com/mozilla-rally/rally-study-01#readme"
+  "homepage": "https://github.com/mozilla-rally/rally-web-platform1#readme"
 }
